requirements.txt
asgiref==3.7.1
blosc2==2.0.0
Cython==0.29.34
Django==4.2.1
django-tables2==2.5.3
djangorestframework==3.14.0
geopy==2.3.0
msgpack==1.0.5
numexpr==2.8.4
numpy==1.24.3
packaging==23.1
psycopg2-binary==2.9.6
py-cpuinfo==9.0.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tables==3.8.0
typing_extensions==4.6.1
